rename tests files to remove wrapper
