+++ tests_pytest/_fw_tests_common_base/fusing/base_graph_with_fusing_metadata_test.py
