+++ tests_pytest/keras_tests/integration_tests/core/fusion/test_graph_with_fusing_metadata_keras.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 
 
-from tests_pytest._fw_tests_common_base.fusing.base_fusion_metadata_wrapper_test import BaseGraphWithFusingMetadataTest
+from tests_pytest._fw_tests_common_base.fusing.base_graph_with_fusing_metadata_test import BaseGraphWithFusingMetadataTest
 from tests_pytest.keras_tests.keras_test_util.keras_test_mixin import KerasFwMixin
 
 import keras
